# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TokenTimer Core API Contract
  version: 0.1.0
  description: |
    TokenTimer Core API for authentication, workspaces, token lifecycle management,
    alerts, audit, and integration import/scan workflows.
    This contract is also used as a compatibility baseline across core, cloud, and enterprise variants.
servers:
  - url: http://localhost:4000
    description: Local development backend
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/workspaces/{id}/contacts:
    get:
      tags: [Workspaces]
      summary: List workspace contacts
      operationId: listWorkspaceContacts
      description: Returns contact records for the workspace.
      parameters:
        - $ref: "#/components/parameters/workspaceIdParam"
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Workspaces]
      summary: Create workspace contact
      operationId: createWorkspaceContact
      description: Creates a workspace contact with email and/or phone details.
      parameters:
        - $ref: "#/components/parameters/workspaceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [first_name, last_name]
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                phone_e164:
                  type: string
                details:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{id}/contacts/{contactId}:
    put:
      tags: [Workspaces]
      summary: Update workspace contact
      operationId: updateWorkspaceContact
      description: Updates mutable fields of a workspace contact.
      parameters:
        - $ref: "#/components/parameters/workspaceIdParam"
        - in: path
          name: contactId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Workspaces]
      summary: Delete workspace contact
      operationId: deleteWorkspaceContact
      description: Deletes a contact from the workspace.
      parameters:
        - $ref: "#/components/parameters/workspaceIdParam"
        - in: path
          name: contactId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Contact deleted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session cookie authentication used by authenticated API routes.
  parameters:
    workspaceIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          oneOf:
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true
            - type: string
      example:
        error: Invalid request payload
        code: VALIDATION_ERROR
        details:
          - email is required
...